medication-tracker/packages/backend/requirements.txt
flask>=2.3.3
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.5
flask-migrate>=4.0.4
flask-jwt-extended>=4.5.2
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
pydantic>=2.1.1
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.5.1
flake8>=6.1.0
black>=23.7.0
isort>=5.12.0
python-jose>=3.3.0
requests>=2.31.0
celery>=5.3.1
redis>=5.0.0
gunicorn>=21.2.0
